+++ docs/source/releasenotes.rst
@@ -5,6 +5,18 @@ Release notes
 `Upcoming release <https://github.com/robocorp/rpaframework/projects/3#column-16713994>`_
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+17.0.0 - 20 Sep 2022
+--------------------
+
+- Library **RPA.Cloud.AWS** (:pr:`637`)
+
+  - All references to `Robocloud.Vault` changed to `Robocorp.Vault` (parameters and documentation).
+    This is **breaking** for this library, which leads to **major** version bump also for ``rpaframework``, 
+    because this library can be also installed with ``rpaframework[aws]`` instead of ``rpaframework-aws==5.0.0``.
+  - Add possibility to pass extra parameters for some S3 keywords
+
+- Library **RPA.Cloud.Google** (:pr:`619`): Add support for `Document AI <https://cloud.google.com/document-ai/>`_  service
+- Library **RPA.Excel.Files** (:pr:`638`): Add support for opening .xlsx files in ``read_only`` mode
 
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++